+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add `tabIndex` to BaseProps
 - Add `Like`, `Dislike`, `Feedback` icons
+- Add `isSelected` prop to NavBarItem
 - Add optional `icon` property for `Radio` and `RadioGroup` to allow for rendering an icon instead of a label
 
 ### Changed